Update HTTP binding protocol generation (#465)
This commit updates the support for aws.rest-json-1.1 to build atop
the refactor of protocol generation done in smithy-typescript. It
also introduces support for the aws.rest-json-1.0 protocol.

In doing so, general components were built atop the abstractions
provided. The aws.rest-json-1.* set of protocols serialize BigInteger
and BigDecimal shapes as JSON numbers on the wire, so an implementation
of the DocumentMemberSerVisitor was created to fail generation.
Implementations of the [DocumentShape[Deser|Ser]Visitor classes are
also included to support serde of shapes in request and response
bodies.

A rest-json implementation of the HttpBindingProtocolGenerator has been
created to combine the above aspects and handling of document bodies.