Add support for AWS rest-xml protocol
This commit adds support for the `aws.rest-xml` protocol, building
on top of the `HttpBindingProtocolGenerator` for document and
payload serde.

Implementations of the `DocumentMember[Deser|Ser]Visitor` and the
`DocumentShape[Deser|Ser]Visitor` have been created that handle
Smithy's XML traits and their influence on protocol serde.

A minor update has been made to the `XmlNode` to allow for nodes to
be renamed, as the same structure may change XML node names when it
is bound to different locations.