Fix multiple XML collection deserialization bugs
This commit fixes a couple issues with generating code to deserialize
structure, union, or body members that target collection members.

1. Unflattened shapes that could return multiple entries in their
collection but only return one would result in an error at runtime
trying to iterate over an object like it was an Array.
2. Unflattened shapes that were not set in the provided output would
result in an error at runtime when trying to validate the entry exists
element without validating that its parent element existed.

This update also bumps the version of smithy-aws-traits to latest.