Add support for the aws.ec2 protocol
This commit adds support for the `aws.ec2` protocol, building on top
of the `HttpRpcProtocolGenerator`, `Xml[Member|Shape]DeserVisitor`,
and `Query[Member|Shape]SerVisitor` for document serde.

The `QueryShapeSerVisitor` has been opened up for re-use by the new
`Ec2ShapeSerVisitor` because `aws.ec2` is a specific version of the
`aws.query` protocol. Hooks have been made available to influence
the specific behavior that is updated.

This also fixes a critical bug with query list and map serialization
that would have resulted in runtime failues on serilaizing the member
target shape's contents.