feat: protocol test codegen project + service clients (#1058)
* Add project for protocol test codegen

This commit adds a separate Gradle project that depends on the
published version of the smithy-aws-protocol-tests. This new project
contains 1 projection for each protocol, generating the full AWS SDK
for JS V3 client and test suite for each contained service.

* Add generated protocol tests

This commit adds the protocol tests generated by the protocol-test-codegen
project to the repository. It enables their building and running through
the use of new package level commands, build:protocols and test:protocols.