feat: update for Smithy 1.0.0
This commit updates the smithy-aws-typescript-codegen package to version 0.2.0.

This commit updates to the GA version of Smithy tooling, including the 0.2.0
version of the smithy-typescript build plugin. It includes necessary updates
to handle changes in protocol traits, streaming components, and other
assorted breaks.

Fixes are also included for several formatting issues.