chore: add fake credentials and regions to unit tests (#1513)
Some functional tests, unit tests and protocol tests constructs new clients
in test suites. If these tests run in environment without any AWS configs,
they will fail with region is missing or cannot resolve credentials. This
change add fake regions and credentials to this test so SDK won't try to
resolve them in the test. It enables the test:all script to be runnable on
for example, a fresh EC2 instance