chore(middleware-host-header): move to low priority
AWS Signer requires a request to contain a `host` header, mostly the same as the request's `hostname`. 
For the customizations that modifies the request's `hostname`, they need to be applied before this 
middleware, Otherwise the `hostname` and `host` header would mismatch. It mean these middleware
**must** be applied relatively before this middleware. By moving it low priority, it reduce the possibility 
of mistakenly adding customizations applied after this middleware. Developers only need to apply the
customizations to the `build` step, and they will be applied before this middleware by default.