fix(generate-clients): Invoke prettier relative to client-generation
It allows to run `scripts/generate-clients` outside of the aws-sdk root directory.

Before it was required to run the script always in the root (like with `yarn generate-clients`. By invoking prettier with relative path it can be executed from every directory like `node $HOME/aws-sdk-js-v3/scripts/generate-clients` which is helpful for building private clients.