chore: merge ssdk branch into main (#2279)
* chore: serialize rest json error code header (#2166)

* chore: short-circuit ssdk-incompatible integrations (#2167)

* chore: disable idempotency autofill import when not generating a client (#2181)

* Support generating non-AWS client (#2222)

* feat(config-resolver): make region optional for non-AWS client

* feat(codegen): skip integrations that are not relevant for non-AWS services

This is an initial version to get a working version of generated code
that compiles without manual edits in smithy-typescript-ssdk-demo. I
expect to make updates to this logic.

* chore(codegen): address code comments

The minor ones from https://github.com/adamthom-amzn/aws-sdk-js-v3/pull/1

* fix(codegen): use SigV4Trait check instead of ServiceTrait

AddAwsRuntimeConfigTest is checking for some behaviors from
AddAwsAuthPlugin too, which was failing with missing aws.auth#sigv4
trait after my change. Added the trait for now to the test, but unit
tests will need to be added/refactored for all these changes.

* chore(codegen): move isAwsService check to utils class

* chore(codegen): code style formatting

* chore(codegen): check SigV4 trait for including region

* chore(codegen,config-resolver): refactor how endpoint is resolved for non-AWS client (#2287)

* chore(config-resolver): refactor EndpointsConfig for non AWS services

This reverts an earlier change to EndpointsConfig.ts and instead
provides the new functionality in separate
CustomEndpointsConfig.ts. This will be used as a separate plugin for
endpoint resolution in non AWS clients.

* feat(codegen): Use separate CustomEndpointsConfig for non-AWS clients

This depends on newly added CustomEndpointsConfig in
@aws-sdk/config-resolver package.

* test(config-resolver): add test for CustomEndpointsConfig (#2305)

Co-authored-by: Adam Thomas <61852529+adamthom-amzn@users.noreply.github.com>
Co-authored-by: Jaykumar Gosar <5666661+gosar@users.noreply.github.com>