fix(endpoint-cache): reduce bundle size (#2400)
mnemonist is a pretty big commonjs package, which is currently included
IN FULL from the endpoint-cache package. This is less than ideal when
using the sdk on the frontend side, since it adds about 270kb to the
bundle (before minification).

Since only the LRU-Cache is needed, changing the import to directly
target the lru-cache module helps bundlers tree shake correctly,
reducing the included code to about 15kb.