feat: support Sigv4 for non AWS services (#2385)
* feat(middleware-signing): support SigV4 for non AWS services

There is no regionInfo to determine signingService. The signingName
will instead directly come from input. And region will be used as
signingRegion.

* feat(codegen): support SigV4 for non AWS services

Write `signingName` from the SigV4 trait for non AWS service.
`region` is used for the signingRegion.
A separate SigV4Auth is used for SigV4 logic for non AWS services.

* fix(codegen): use Region plugin explicitly for AWS service

The models for the protocol tests don't seem to have the SigV4 trait,
so can't rely solely on it for using the Region plugin. e.g,
https://github.com/awslabs/smithy/blob/main/smithy-aws-protocol-tests/model/restJson1/main.smithy#L11-L13

* chore(clients): update clients with updates for SigV4 for non AWS services

* test(middleware-signing): add test for non AWS service SigV4

* chore(codegen): reduce duplication for writing region

* Reverting unnecessary changes to protocol tests