chore(codegen): generate and run SSDK protocol tests during CI (#2558)
This sets up a server-only mode for generate-clients, and a set of yarn targets
that generate and run server protocol tests. The CodeBuild CI for clients uses
Node 10, while SSDKs require Node 14, so this sets up an additional GitHub
action for running the SSDK protocol tests and excludes SSDK protocol tests
from the client targets. When the client is updated to support 14 as a minimum
version, this distinction can be removed.