feat(middleware-api-key): add new auth middleware
We will create a new middleware to handle API keys in the
context of Smithy generated models (probably not
applicable for AWS specific services). The code is based
heavily on [this PR](
https://github.com/awslabs/smithy-typescript/pull/473
) from the Smithy TypeScript repository.

In order for it to work, we need to ensure the Bearer token
middleware succeeds even when the `token` is not set.