fix(codeowners): fix syntax and precedence (#4378)
The CODEOWNERS file was set up with incorrect syntax and
precedence rules. Previously, the syntax meant that smithy
were owners on only files in the root of each specified
directory, but not any subdirectories. Additionally, the
precedence rules meant that Smithy would be codeowners on
these files, rather than both the Smithy team and JS team.
This PR fixes both issues.