feat(protocol-http): implement SRA HttpRequest
This change implements the core functionality of HttpRequest for SRA.
New properties were added to the HttpRequest class, old properties
and interfaces were deprecated. Proxy was added for headers and query
properties to maintain backward compatibility. This change does not
include updates to the SDK to use new functionality.

Notes on other changes:
* middleware-host-header and middleware-sdk-transcribe-streaming tests:
  URL.port always returns '' if port is set to the default for the
protocol.
* middleware-sdk-transcribe-streaming: Providing URL.hostname with a
  string containing the port number does not update the port.
* signature-v4 prepareRequest: See comment in code.
* Field: toString method was changed to only require double quotes on
  values with commas. Current tests allow spaces without double quotes.
* Fields: Helper methods were added for getting all fields and creating
  a Fields object.

Additional notes:
* When converting between single string header values and a list of
  field values, split(",") and join(",") are used to avoid making any
modifications to the original value, preserving compatibility.
* HttpRequest.isInstance was not modified, because doing so might break
  compatibility. The check was not looking for the clone method, so
wasn't actually verifying the object was indeed an instance. The method
could easily have been used on a HttpRequest interface from the types
package, and returned true.