feat(protocol-http): implement SRA HttpResponse
This change implements the core functionality of HttpResponse for SRA.
New properties were added to the HttpResponse class, old properties and
interfaces were deprecated. To maintain compatibility with old
implementation, Proxy was added for headers property, and
Object.defineProperties is used with getters/setters for other old
properties. This change does not include updates to the SDK to use new
functionality.

Additional notes:
* HttpResponse.isInstance was left unchanged to maintain compatibility.
  Previous implementation only checked for specific properties, so it
would return true when passed the HttpResponse interface from types
package.
* https://github.com/awslabs/smithy-typescript/pull/698 is a related
  change that ensures protocol tests are using HttpResponse constructor.
Due to type changes in HttpResponse, this is required.