fix: enable keep-alive in node-http-handler
When using node-http-handler and setting keep-alive to true in the agent configuration, this setup is not being honored by nodejs https library due to a bug at their side. To aid this issue at the SDK side what I am doing is enabling keep alive directly in the socket, by adding a listener for when a new socket is created, and there we provide the keep-alive settings to the socket. I add this configuration as option so that we need to explicity say we want to enable keep-alive in the socket to keep forward compatibility.