test(client-s3): add path encoding test (#4574)
The S3 client uses a configuration option, `signingEscapePath`, set
to false by default, to tell the signer not to encode the URL path.
The serializer will encode these paths, so if the signer did as well,
the paths would be encoded multiple times. S3 is the only client that
uses this config option, and there were no tests that verified paths
are encoded.

This is a followup to #4523.