feat(protocol-http): add reason to HttpResponse
Adds `reason` property to HttpResponse, which is a simple message
explaining the status code. The property is optional, so adding it
is backward compatible. Http handlers were updated to populate this
property when constructing the HttpResponse, except for the Node http2
handler, because http2 doesn't support an equivalent.

Tests were added to Http handlers to make sure the property is populated
when available.