chore(codegen): fix codegen release
Jreleaser has to work in a git repository, and uses the jgit java
library to interact with git. When jgit tries to open the git repo,
it tries to open `codegen/config` as a file, which blows up. I don't
know of any way to configure what jgit is doing here, so I renamed
the directory and it solves the issue. Also had to set `gitRootSearch`
so jreleaser walks up the parent directories to find the git root.